Use 1€ filter for landmarks, try selecting wide angle camera

+++ app/src/main/java/com/filippoorru/topout/model/State.kt
@@ -9,7 +9,7 @@ enum class ClimbingState {
 }
 
 data class PoseState(
-    val feet: List<Pair<Float, Float>>,
+    val feet: List<Pair<Double, Double>>,
     val feetTrackingPoints: List<TrackingPoint>,
     val averageDuration: Long,
 )
